[back] refactor : 사용자 스크랩 브랜드 편집 기능 수정

+++ src/main/java/com/pard/server/fashion_muse/user/service/UserService.java
@@ -33,23 +33,25 @@ public class UserService {
     }
 
     @Transactional
-    public void deleteUserScrap(Long userId, Long brandId) {
+    public void deleteUserScrap(Long userId, List<Long> brandIds) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("유저가 존재하지 않습니다."));
 
-        Brand brand = brandRepository.findById(brandId)
-                .orElseThrow(() -> new RuntimeException("브랜드가 존재하지 않습니다."));
+        List<Brand> brands = brandRepository.findAllById(brandIds);
 
-        UserScrap userScrap = userScrapRepository.findByUserAndBrand(user, brand)
-                .orElseThrow(() -> new RuntimeException("해당 스크랩이 존재하지 않습니다."));
+        for (Brand brand : brands) {
+            UserScrap userScrap = userScrapRepository.findByUserAndBrand(user, brand)
+                    .orElseThrow(() -> new RuntimeException("스크랩 완료된 유저의 브랜드가 존재하지 않습니다."));
 
-        userScrapRepository.delete(userScrap);
-
-        brand.setScrapCount(
-                brand.getScrapCount() == null || brand.getScrapCount() <= 0
-                        ? 0
-                        : brand.getScrapCount() - 1
-        );
+            if (userScrap != null) {
+                userScrapRepository.delete(userScrap);
+                brand.setScrapCount(
+                        brand.getScrapCount() == null || brand.getScrapCount() <= 0
+                                ? 0
+                                : brand.getScrapCount() - 1
+                );
+            }
+        }
     }
 
 }
